Satellite imagery projects/water-detect-web/backend/requirements.txt
flask
ultralytics
pillow
simplekml
shapely
requests
